Added a picture
Added a picture and changed constants from int to float

+++ src/main/java/ch/bbcag/jamkart/Constants.java
@@ -4,11 +4,11 @@ public class Constants {
 
     public static final float NETWORK_TICK_TIME = 0.1f;
 
-    public static final int START_WINDOW_WIDTH = 800;
-    public static final int START_WINDOW_HEIGHT = 600;
-    public static final int GAME_WINDOW_WIDTH = 1500;
-    public static final int GAME_WINDOW_HEIGHT = 800;
+    public static final float START_WINDOW_WIDTH = 800.0f;
+    public static final float START_WINDOW_HEIGHT = 600.0f;
+    public static final float GAME_WINDOW_WIDTH = 1500.0f;
+    public static final float GAME_WINDOW_HEIGHT = 800.0f;
 
-    public static final int CAR_SPAWN_FIRST_Y = 65;
-    public static final int CAR_SPAWN_DISTANCE = 50;
+    public static final float CAR_SPAWN_FIRST_Y = 65.0f;
+    public static final float CAR_SPAWN_DISTANCE = 50.0f;
 }
